Start off with a simple example using built-in resources
We're jumping into complex resources that requiring loading the current value, and that's going to scare off users. Custom resources using built in chef resources handle idempotency for the user and are dead simple to write. We should introduce the concept of loading the current value and converging on that value later.

Signed-off-by: Tim Smith <tsmith@chef.io>