Add the property validation methods
Properties have some validation methods that were not documented.  Added and
modified the documentation of the validation methods from pre 12.5.1
resource attributes.  Documentation for the missing methods was too hard to find unless
you knew it existed. It was also not obvious that the methods had been carried forward
with the new style custom resources.