Removed start and end dtags
All docs under `/snapshots` are still built using multiple include files and do not reflect the recent change to the dtag syntax used in the `chef_master` directory. Therefore, we don't need to add in this tag. After we flatten these snapshot dirs, then new tags will be created for any shared content as part of the flattening process.