Make Recipe DSL reachable in the nav.
This solution makes Recipe DSL a sibling of Recipe rather than a child.
The heirarchy is altered, but the fine grain links are retained. This
appears to have been davidwrede's intent according to
https://github.com/chef/chef-web-docs/pull/143

NOTE: To make such a significant heirarchy change I had to lint the
JSON. This resulted in a lot of minor whitespace and comma corrections.
I wasn't trying to be pedantic.