Minor tweaks to the getting started page
There's no need to talk about chef-zero with new users. It's a backend
tech that enables cool tooling, but itself isn't the cool tooling. We
should also remove references to definitions since no one new to Chef
needs to know about these. Finally ChefDK comes with Cookstyle, but
"Policy" isn't really a tool. It's a Chef feature and it seems out of
place in this list.

Signed-off-by: Tim Smith <tsmith@chef.io>