[chef-backend] New defaults for etcd and leaderl tunables
Chef Backend 1.2 will ship with different default values for a few
etcd and leaderl tunables. This change updates the documentation to
reflect the new defaults.

In the sections I modified, I allowed my editor to auto-fill the
paragraphs.  This doesn't affect the final rendering as far as I can
see, but does make it much easier to read in an editor.

Signed-off-by: Steven Danna <steve@chef.io>