Rework getting started for 2016 (#292)
* This is confusing in a bulleted list

It’s just not needed here

Signed-off-by: Tim Smith <tsmith@chef.io>

* Chefspec doesn’t run the chef-client

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove chef-zero from the list of tools included

It’s included, but there’s really no reason to mention it in a getting started document. It’s just powering the tools that people actually care about

Signed-off-by: Tim Smith <tsmith@chef.io>

* ChefDK included the Chef CLI not Chef itself

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove definitions from the overview

We document them later, but it’s really not something we should expose new users to

Signed-off-by: Tim Smith <tsmith@chef.io>

* A workstation == ChefDK

We need to stress this more. Also stress testing in ChefDK vs. “command line tools”

Signed-off-by: Tim Smith <tsmith@chef.io>

* Cleanup a very confusing sentence

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update other tags

Signed-off-by: Tim Smith <tsmith@chef.io>

* Test Kitchen is the preferred name

At one point the idea was to call it Kitchen.ci, but it’s referred to as Test Kitchen now and we should be consistent here to avoid confusion.

Signed-off-by: Tim Smith <tsmith@chef.io>

* CherDK is a product. Capitalize it

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix tenses and remove reference to node object

No need to bring up node object *yet*

Signed-off-by: Tim Smith <tsmith@chef.io>

* We don’t author cookbooks on the Supermarket

Signed-off-by: Tim Smith <tsmith@chef.io>

* This note makes no sense for new users

Signed-off-by: Tim Smith <tsmith@chef.io>

* Clarify where settings are made

We’re talking about chefdk in this paragraph not chef-client.

Signed-off-by: Tim Smith <tsmith@chef.io>

* This diagram is components and tools

Chef-repo is not a tool. These are the tools and components of your workstation

Signed-off-by: Tim Smith <tsmith@chef.io>

* Rework the ChefDK tool list

group chef-client/ohai since they’re a pair
clarify we’re talking about the chef cli tool, not the chef-client

Signed-off-by: Tim Smith <tsmith@chef.io>

* Shared is a better word for what Supermarket does

Signed-off-by: Tim Smith <tsmith@chef.io>

* Replicating shared content changes

Signed-off-by: David Wrede <dwrede@chef.io>

* Updates based on review

I nuked most of the description of the workstation’s role. The sentence was a horrible run on and included details that just don’t fit in an introduction sentence.

Signed-off-by: Tim Smith <tsmith@chef.io>