Show that push server 2.1 is now supported with automate
This adds upgrade instructions to the push jobs release notes
and references to them in install/upgrade automate docs.

The documented 1.x automate restrictions has also been removed.

Note that the release notes are not the ideal place for this change, but
it's the best available option that doesn't involve making a new page
that's more difficult to discover.

Signed-off-by: Marc Paradise <marc@chef.io>