Clarify use of --json-attributes
Someone in the #chef IRC channel was confused about use of this option - although they'd read this documentation, they mistook the example as being something that would work for setting override_attributes.  While the docs are accurate, I agree it's easy to misinterpret them if you're not paying close attention.