Rework supported platforms
Add FreeBSD 11 as as a secondary platform as we have a package for it, but we're not entirely testing it yet. Remove Debian 6 which is EOL, but add 9 which we're testing in the cookbook ecosystem now.

We claim that secondary platforms have packages, but we don't test them. Not sure why Arch in Gentoo got in here then since they have no packages.

Signed-off-by: Tim Smith <tsmith@chef.io>