Move knife usage information under the knife doc
The knife doc in its current form contains some information on subcommands, but a lot of the useful information to make that complete exists in the using knife doc. This moves all the basic usage information into the common knife page. It also reworks information on knife plugins by providing instructions for their installation, updating the list of chef maintain plugins, and correctly linking to the supermarket for community maintained plugins.

Signed-off-by: Tim Smith <tsmith@chef.io>