Fix incorrect source example for chef server
I believe the example for getting Berkshelf to work with open source chef is incorrect. I think the syntax should be similar to the local Chef repository example below. If I am wrong let me know and I will withdraw this PR but this is how I had to configure my `Berksfile` on my system to get it working.