Remove documentation for the path property
We removed the path property in Chef 12 which was released at the end of 2014. Chef 11 hasn't been supported for nearly a year and Chef 12 itself is actually going EOL in 3 months. We have a Foodcritic rule to alert users to this coding error. There's no need to pollute the docs with it.