Remove "group" property and remove confusion with manage_home
We had a bug report that manage_home docs were confusing because it seems like manage_home is always set with the :create action. I switched up the words a tiny bit to make it more clear that were were describing the behavior when manage_home was set and you were in the :create / :modify actions. I think this makes it a bit more clear. I also removed the old 'group' property which has been gone for a while. I mention the old name, but there's no reason to define the entire thing. It's an alias and we have tons of these throughout Chef.