Automatically deploy using Workflow following merges
This removes the manual `delivery review` step that was previously
needed to deploy changes to acceptance. Now each time a PR merge happens
on GitHub, Expeditor will then trigger the proper Workflow pipeline and
automatically deploy changes to acceptance.

The auto-deploy to Workflow can be skipped by applying one of the
following labels to the GitHub PR:

- `ACC: Skip Update`
- `Expeditor: Skip All`

These are configured in the `.expeditor/config.yml` file. The
`ignore_labels` action filter is documented here:
https://expeditor.chef.io/docs/reference/action-filters/#ignore-labels

Signed-off-by: Seth Chisamore <schisamo@chef.io>