Update FreeBSD 10 EOL date, update SLES support, remove .x versions (#1516)
* Remove FreeBSD 10, update SLES, remove .x versions

FreeBSD 10 went EOL April 30th according to our docs, but it was still in the table. This removes it. It also clarified we support SLES 12 SP1 or later. We had that in a few places, but not all. Finally it simplifies how we refer to major release versions.

* dtags fix

* Add back FreeBSD 10 with updated EOL

Upstream has changed the date for 10 going EOL

Signed-off-by: Tim Smith <tsmith@chef.io>