Stop applying redirects...for now
This process currently takes over an hour to complete and our redirects
rarely change. The @chef/docs-team is working to curate this list of
redirects and remove old crufty items. Once this work is done we can
re-enable the processing of the list or better yet: move this logic to
the Sphinx code.

Signed-off-by: Seth Chisamore <schisamo@chef.io>