Update how we suggest users share cookbooks
In Chef 13+ we ship both the knife cookbook site and knife supermarket commands. The cookbook site was the pre-2014 site and we decided that in Chef 15 we're going to deprecate the old cookbook site commands. The first step is updating our docs to suggest everyone use the supermarket commands.

Also fixes a few inaccurate statements about Supermarket

Signed-off-by: Tim Smith <tsmith@chef.io>