Remove references to definitions
This is one of those features that causes people to have a really bad time with Chef. We've been warning for many years to not use this feature in Foodcritic and we've had a warning on the docs site since mid Chef 12. This is the next step in our deprecation of this feature: Hide it unless someone has it bookmarked.

In the near future we'll begin throwing formal deprecation warnings in the chef-client and then we'll remove the feature altogether.

Signed-off-by: Tim Smith <tsmith@chef.io>