Update resource_package so not installing in loop.
Installing packages in a loop, like this:

```
   package %w{package-a package-b package-c package-d} do
     action :upgrade
   end
```

Causes unique package resources to be generated.  This increases the overall node object size and causes node runs to take much longer (O*n).  As such, I would argue that defining a list of packages then iterating over that list is an anti-pattern, especially considering the package resource takes an array.

Additionally, this is false:

    and the next uses a single **package** resource and a whitespace array (``%w``):

As iterating over the loop dynamically creates package resources equal to the number of packages.

Perhaps this would be more clear?

    pkgs = %w{package-a package-b package-c package-d}
    package pkgs do
      action :upgrade
    end

Thanks,
-Q