Remove config information for Audit mode
This was a beta feature that never came out of beta and is being removed
from the client entirely. No one needs to know about these config
options and it only confuses users since we have a modern "audit" mode
via the audit cookbook, which is an entirely different thing.

Signed-off-by: Tim Smith <tsmith@chef.io>