More resource docs cleanup (#1752)
* Add missing properties to windows_task resource

Due to a bug in the resource it wasn't showing up with the resource
inspector before. This is fixed so now we get to fix some issues.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove carriage returns from docs

This make our dtags different in hidden ways and it's super hard to
track down without enabling hidden characters

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix new in chef client formatting

Signed-off-by: Tim Smith <tsmith@chef.io>

* Bring in a few more changes from automation

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove extra whitespace

Signed-off-by: Tim Smith <tsmith@chef.io>

* Don't mention very old Chef 12 releases

I left in more recent notes about Chef 12, but these are just ancient and we shouldn't litter our notes with information that isn't relevant anymore.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Reporting -> Automate in the config docs

Signed-off-by: Tim Smith <tsmith@chef.io>

* Squash a "new in" for Chef 13

Just explain the behavior. If someone is upgrading from Chef 12 this is called out in our docs

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove 2 unused configs

knife syntax checking was removed.
yum has been refactored to not honor this config anymore

Signed-off-by: Tim Smith <tsmith@chef.io>

* dtag fixes

Signed-off-by: Tim Smith <tsmith@chef.io>