Add more detailed general recovery process
We have a lot of separate scenarios here. Let's try to make a unified one, then refer to the separate scenarios from it.
A common scenario are followers that have fallen so far out of sync, the `chef-server-ctl join-cluster IPADDRESS --recovery` command no longer works to rejoin a follower to the cluster.