Remove docs for the EOL Audit Mode (#2002)
* Remove docs for the EOL Audit Mode

We introduced this as an experimental feature in 2015. It has to be turned on via the config. Shortly after that we released InSpec and this feature was never worked on after that or really used by anyone. We removed it in DK 4 / Chef 15. We should entirely wipe it from our memories.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove some older / broken knife plugins from our list (#2003)

Remove the Chef written plugins that we don't maintain anymore
Remove knife-voxel as the company is no more
Remove knife-community as knife supermarket is built into chef now
Remove knife-baremetalcloud as it hasn't been updated in 7 years

Signed-off-by: Tim Smith <tsmith@chef.io>

* Remove dsl_audit from index.rst

Signed-off-by: IanMadd <imaddaus@chef.io>