Remove the Chef Client and Knife plugins
Chef Client doesn't actually have plugins. This was just an odd list of handlers and some random tools that don't make sense in 2019.

The knife plugins in this list are of real limited value. Most of these are not maintained and I would guess many no longer work. We really shouldn’t list these here since we have this on the Supermarket now.

Signed-off-by: Tim Smith <tsmith@chef.io>