Update the Test Kitchen docs (#2012)
* .kitchen.yml -> kitchen.yml

This is the modern name of the config and the name used by all our generators.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Modernize our examples

New Test Kitchen
Chef 15
Ubuntu 18.04
InSpec
Proper indentation

Signed-off-by: Tim Smith <tsmith@chef.io>

* Kitchen -> Test Kitchen

The product is Test Kitchen and the command is kitchen. When we're using big K kitchen we should be calling it Test Kitchen.

Signed-off-by: Tim Smith <tsmith@chef.io>

* DK -> Workstation

Describe the path we want folks to be on

Signed-off-by: Tim Smith <tsmith@chef.io>

* Kitchen isn't used by Chef Infra

This was an odd statement. I hope this is less odd.

Signed-off-by: Tim Smith <tsmith@chef.io>

* 1404 -> 1804 and fix typos

The output here wasn't correct and that was probably confusing folks

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix a few more branding issues

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update to a more modern version

Signed-off-by: Tim Smith <tsmith@chef.io>

* Better describe what Test Kitchen is

The architecture of the plugin system isn't that important
Make sure people understand that Bento is our project
Mention Policyfiles

Signed-off-by: Tim Smith <tsmith@chef.io>

* Rework Plugins and Validation

Simplify the validation section to just mention what's possible. We might want to just link to a learnchef module here

Talk about modern drivers that people would actually care about

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update the list of common drivers

Remove things that are unmaintained at this point

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update the things we suggest folks read

We don't want new folks jumping on Berkshelf

Signed-off-by: Tim Smith <tsmith@chef.io>

* Kitchen -> Test Kitchen

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update chef_master/source/kitchen.rst

Signed-off-by: Tim Smith <tsmith@chef.io>

Co-Authored-By: mjingle <mjingle@users.noreply.github.com>

* Update chef_master/source/kitchen.rst

Signed-off-by: Tim Smith <tsmith@chef.io>

Co-Authored-By: mjingle <mjingle@users.noreply.github.com>

* dtag fixes

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix a few more dtags

Signed-off-by: Tim Smith <tsmith@chef.io>

* formatting and dtag fixes

Signed-off-by: Mary Jinglewski <mjinglewski@chef.io>