Clarify the use of properties in load_current_value
Show how to mark properties so they will be available in a load_current_value block.
Add name_property to this list of settings that exposes a property.
Make it clearer that the block argument supplied to load_current_value has the values of the new_resource.
Verified that the load_current_value examples work.

Signed-off-by: markgibbons <mark.gibbons@nordstrom.com>