More changes to environments
Explain cookbook version pinning

Remove Attribute Persistence which should only be in the attributes doc

Remove the () from node.environment. This isn't javascript

Move the chef solo instructions to their own category

Make an initial mention of cookbook pinning. This is a super important feature that all our large customers do and we don't even call it out here. Kinda weird

Remove the odd section explaining what Ruby is. We have that earlier on and we don't have a similar section for what JSON is. Eventually we need to get hover overs or aggressive links for this kinda thing. Right now though this just bloats the doc and detracts from the real content.

Signed-off-by: Tim Smith <tsmith@chef.io>