More changes to environments (#2144)
* More changes to environments

Explain cookbook version pinning

Remove Attribute Persistence which should only be in the attributes doc

Remove the () from node.environment. This isn't javascript

Move the chef solo instructions to their own category

Make an initial mention of cookbook pinning. This is a super important feature that all our large customers do and we don't even call it out here. Kinda weird

Remove the odd section explaining what Ruby is. We have that earlier on and we don't have a similar section for what JSON is. Eventually we need to get hover overs or aggressive links for this kinda thing. Right now though this just bloats the doc and detracts from the real content.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Add better version contraint examples

Signed-off-by: Tim Smith <tsmith@chef.io>

* Cookstyle fixes to the examples

Signed-off-by: Tim Smith <tsmith@chef.io>

* Replicate dtag

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update chef_master/source/environments.rst

Co-Authored-By: Kimberly Garmoe <kgarmoe@chef.io>

Signed-off-by: Tim Smith <tsmith@chef.io>
Sign-off-executed-by: Tim Smith <tsmith84@gmail.com>
Approved-at: behalf

* Update chef_master/source/environments.rst

Co-Authored-By: Kimberly Garmoe <kgarmoe@chef.io>

Signed-off-by: Tim Smith <tsmith@chef.io>
Sign-off-executed-by: Tim Smith <tsmith84@gmail.com>
Approved-at: behalf

* Update chef_master/source/environments.rst

Co-Authored-By: Kimberly Garmoe <kgarmoe@chef.io>

Signed-off-by: Tim Smith <tsmith@chef.io>
Sign-off-executed-by: Tim Smith <tsmith84@gmail.com>
Approved-at: behalf

* Update the environment description to actually describe what it does

Move the _default junk elsewhere and focus on what an environment actually does for a user.

Signed-off-by: Tim Smith <tsmith@chef.io>

* More description updates

Signed-off-by: Tim Smith <tsmith@chef.io>