Adding explicit documentation that the link resource creates junction… (#2213)
* Adding explicit documentation that the link resource creates junctions/reparse points on Windows

The link resource can create junctions/reparse points on Windows, but the documentation doesn't explicitly state that. I had to do some experimentation to figure it out. I'd like others to know this without having to do the experimentation.

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the
    best of my knowledge, is covered under an appropriate open
    source license and I have the right under that license to
    submit that work with modifications, whether created in whole
    or in part by me, under the same open source license (unless
    I am permitted to submit under a different license), as
    Indicated in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including
    all personal information I submit with it, including my
    sign-off) is maintained indefinitely and may be redistributed
    consistent with this project or the open source license(s)
    involved.

Signed-off-by: Aaron Jensen <ajensen@webmd.net>

* dtags and formatting fix

Signed-off-by: kagarmoe <kgarmoe@chef.io>

Co-authored-by: Kimberly Garmoe <kgarmoe@chef.io>