RFC2606 - example.com
I also think it's less confusing when using the example domains in documentation rather than specific domain names. This way people can see to fill in their own values rather than a specific hostname that may or may not exist, or that may or may not have permission to use.