Rework the story we're telling in the validatorless bootstrap section (#2528)
* Rework the story we're telling in the validatorless bootstrap section

When we originally wrote this the feature was new so we were letting folks know how to migrate. It's not new at this point and we warn folks not to use it when they bootstrap. Instead we should tell them *why* they shouldn't use it and show them how to move off it. I nuked the layout file that was only used in this single place. There's a ws_ version of it that's not used anywhere that should also go.

Signed-off-by: Tim Smith <tsmith@chef.io>

* validatorless revisions

Signed-off-by: kagarmoe <kgarmoe@chef.io>

* typos

Signed-off-by: kagarmoe <kgarmoe@chef.io>

* Update content/install_bootstrap.md

Signed-off-by: Tim Smith <tsmith@chef.io>

Co-authored-by: Kimberly Garmoe <kgarmoe@chef.io>

* Update content/install_bootstrap.md

Signed-off-by: Tim Smith <tsmith@chef.io>

Co-authored-by: Kimberly Garmoe <kgarmoe@chef.io>

Co-authored-by: kagarmoe <kgarmoe@chef.io>