Move chef-backend.rb settings to the common place
Having the Chef Backend settings deep under Setup -> Chef Infra Server
-> Install High Availability, but not anywhere under Managing Chef Infra
Server means I can never find them.

All of the -ctl commands for Infra Server are under 'Managing Chef Infra
Server' as is the chef-server.rb settings. This moves the the settings
to the same location and links back. These rarely need to be modified,
so they really should not be all spelled out at length on that page.

There's a slight reordering of the menu to group the server, backend,
and expander options in it.

Signed-off-by: Bryan McLellan <btm@loftninjas.org>