Improve some Windows Infra Client docs
Link to the actual downloads page
Remove comparisons to running Infra Client as a service. We're not selling them on tasks anymore. We're just telling them how it is
Link to the chef_client_windows_task resource
Remove the old windows cookbook link. This is EOL and not needed in modern chef
Fix resource names to be lowercase and fix a few typos

Signed-off-by: Tim Smith <tsmith@chef.io>