Create a pipeline for triggering during head of channel promotions
For projects that use the "head of channel" promotion pattern (e.g.
Automate and Habitat) we cannot leverage subscriptions configured in
this repo as many of the promotions tasks that are executed in said
promotions will not be complete when our bash script is executed (think
updating project manifests).

In order to work around this we will create a new pipeline which can be
triggered (using Expeditor's `trigger_pipeline`) as one of the actions
in the `project_promoted` workload subscription in the actual projects
being promoted. Although it's not ideal that our projects which use the
"head of channel" promotion pattern will need to know about
`chef/chef-web-docs` this approach avoids nasty workaround like `sleep`
statements in our bash script.

Signed-off-by: Seth Chisamore <schisamo@chef.io>