Fix malformed warnings in two resources
This is the same bug as in #3521 that I fixed in #3522 but on two other doc pages.
These doc pages render an empty Warning block because the YAML contains an array of hashes
instead of a hash. This should be a hash with a key of `markdown` not an array containing an element that's a hash with a key of `markdown`.

Signed-off-by: Gene Wood <gene_wood@cementhorizon.com>