update quiet mode documentation to reflect actual behavior (#4373)
* update quiet mode documentation to reflect current behavior

this is related to the change that we did to fix quiet mode in https://github.com/chef/chef/pull/14779. in brief, quiet mode wasn't working before. so with the fix we didn't want to break anyone's cofiguration. so quiet mode was changed to be set to false by default. the documentation was also a little confusing since quiet mode is supposed to suppress the output completely vs "controlling verbosity".

Signed-off-by: rishichawda <rishichawda@users.noreply.github.com>

* Clarify that this works starting in Client 18.7

Signed-off-by: Ian Maddaus <ian.maddaus@progress.com>

---------

Signed-off-by: rishichawda <rishichawda@users.noreply.github.com>
Signed-off-by: Ian Maddaus <ian.maddaus@progress.com>
Co-authored-by: Ian Maddaus <ian.maddaus@progress.com>
Co-authored-by: Ian Maddaus <IanMadd@users.noreply.github.com>